refactor: ページ内インラインコンポーネントを分離 (#113)
page.tsx内にインライン定義されていたServer Componentを
featuresディレクトリに分離。

- TitlePageContent → features/title/components/title-page-content/
- ExplorePageContent → features/explore/components/explore-page-content/
- ConnectionPageContent → features/connection/components/connection-page-content/

Note: Server Componentはindex.tsからエクスポートせず、
page.tsxから直接インポートする（バレルエクスポート問題回避）
