+++ src/features/title/components/title-page-content/TitlePageContent.tsx
@@ -1,0 +1,12 @@
+import { getUser } from "@/features/user/_lib/fetcher";
+import TitlePageClient from "../title-page-client/TitlePageClient";
+
+// Server Componentでユーザー情報を取得するラッパー
+const TitlePageContent = async () => {
+	const user = await getUser();
+	const initialZennUsername = user?.zennUsername ?? null;
+
+	return <TitlePageClient initialZennUsername={initialZennUsername} />;
+};
+
+export default TitlePageContent;
